Resolved pull request review items
