+++ src/uhs/atomizer/sentinel/controller.cpp
@@ -86,7 +86,8 @@ namespace cbdc::sentinel {
 
     auto controller::execute_transaction(transaction::full_tx tx)
         -> std::optional<cbdc::sentinel::execute_response> {
-        const auto res = transaction::validation::check_tx(tx, m_opts);
+        const auto res
+            = transaction::validation::check_tx(tx, m_opts.m_minter_pubkeys);
         tx_status status{tx_status::pending};
         if(res.has_value()) {
             status = tx_status::static_invalid;
@@ -118,7 +119,8 @@ namespace cbdc::sentinel {
 
     auto controller::validate_transaction(transaction::full_tx tx)
         -> std::optional<validate_response> {
-        const auto res = transaction::validation::check_tx(tx, m_opts);
+        const auto res
+            = transaction::validation::check_tx(tx, m_opts.m_minter_pubkeys);
         if(res.has_value()) {
             return std::nullopt;
         }
