+++ tests/unit/validation_test.cpp
@@ -45,14 +45,18 @@ class WalletTxValidationTest : public ::testing::Test {
 };
 
 TEST_F(WalletTxValidationTest, valid) {
-    auto err = cbdc::transaction::validation::check_tx(m_valid_tx, m_opts);
+    auto err
+        = cbdc::transaction::validation::check_tx(m_valid_tx,
+                                                  m_opts.m_minter_pubkeys);
     ASSERT_FALSE(err.has_value());
 }
 
 TEST_F(WalletTxValidationTest, no_inputs) {
     m_valid_tx.m_inputs.clear();
 
-    auto err = cbdc::transaction::validation::check_tx(m_valid_tx, m_opts);
+    auto err
+        = cbdc::transaction::validation::check_tx(m_valid_tx,
+                                                  m_opts.m_minter_pubkeys);
     ASSERT_TRUE(err.has_value());
     ASSERT_TRUE(
         std::holds_alternative<cbdc::transaction::validation::tx_error_code>(
@@ -70,7 +74,9 @@ TEST_F(WalletTxValidationTest, no_inputs) {
 TEST_F(WalletTxValidationTest, no_outputs) {
     m_valid_tx.m_outputs.clear();
 
-    auto err = cbdc::transaction::validation::check_tx(m_valid_tx, m_opts);
+    auto err
+        = cbdc::transaction::validation::check_tx(m_valid_tx,
+                                                  m_opts.m_minter_pubkeys);
     ASSERT_TRUE(err.has_value());
     ASSERT_TRUE(
         std::holds_alternative<cbdc::transaction::validation::tx_error_code>(
@@ -86,7 +92,9 @@ TEST_F(WalletTxValidationTest, no_outputs) {
 TEST_F(WalletTxValidationTest, missing_witness) {
     m_valid_tx.m_witness.clear();
 
-    auto err = cbdc::transaction::validation::check_tx(m_valid_tx, m_opts);
+    auto err
+        = cbdc::transaction::validation::check_tx(m_valid_tx,
+                                                  m_opts.m_minter_pubkeys);
     ASSERT_TRUE(err.has_value());
     ASSERT_TRUE(
         std::holds_alternative<cbdc::transaction::validation::tx_error_code>(
@@ -102,7 +110,9 @@ TEST_F(WalletTxValidationTest, missing_witness) {
 TEST_F(WalletTxValidationTest, zero_output) {
     m_valid_tx.m_outputs[0].m_value = 0;
 
-    auto err = cbdc::transaction::validation::check_tx(m_valid_tx, m_opts);
+    auto err
+        = cbdc::transaction::validation::check_tx(m_valid_tx,
+                                                  m_opts.m_minter_pubkeys);
 
     ASSERT_TRUE(err.has_value());
     ASSERT_TRUE(
@@ -122,7 +132,9 @@ TEST_F(WalletTxValidationTest, duplicate_input) {
     m_valid_tx.m_witness.emplace_back(m_valid_tx.m_witness[0]);
     m_valid_tx.m_outputs[0].m_value *= 2;
 
-    auto err = cbdc::transaction::validation::check_tx(m_valid_tx, m_opts);
+    auto err
+        = cbdc::transaction::validation::check_tx(m_valid_tx,
+                                                  m_opts.m_minter_pubkeys);
 
     ASSERT_TRUE(err.has_value());
     ASSERT_TRUE(
@@ -139,7 +151,9 @@ TEST_F(WalletTxValidationTest, duplicate_input) {
 
 TEST_F(WalletTxValidationTest, invalid_input_prevout) {
     m_valid_tx.m_inputs[0].m_prevout_data.m_value = 0;
-    auto err = cbdc::transaction::validation::check_tx(m_valid_tx, m_opts);
+    auto err
+        = cbdc::transaction::validation::check_tx(m_valid_tx,
+                                                  m_opts.m_minter_pubkeys);
 
     ASSERT_TRUE(err.has_value());
     ASSERT_TRUE(
@@ -156,7 +170,9 @@ TEST_F(WalletTxValidationTest, invalid_input_prevout) {
 
 TEST_F(WalletTxValidationTest, asymmetric_inout_set) {
     m_valid_tx.m_outputs[0].m_value--;
-    auto err = cbdc::transaction::validation::check_tx(m_valid_tx, m_opts);
+    auto err
+        = cbdc::transaction::validation::check_tx(m_valid_tx,
+                                                  m_opts.m_minter_pubkeys);
 
     ASSERT_TRUE(err.has_value());
     ASSERT_TRUE(
@@ -253,7 +269,9 @@ TEST_F(WalletTxValidationTest, witness_p2pk_invalid_signature) {
 TEST_F(WalletTxValidationTest,
        check_transaction_with_unknown_witness_program_type) {
     m_valid_tx.m_witness[0][0] = std::byte(0xFF);
-    auto err = cbdc::transaction::validation::check_tx(m_valid_tx, m_opts);
+    auto err
+        = cbdc::transaction::validation::check_tx(m_valid_tx,
+                                                  m_opts.m_minter_pubkeys);
     ASSERT_TRUE(err.has_value());
     ASSERT_TRUE(
         std::holds_alternative<cbdc::transaction::validation::witness_error>(
@@ -339,72 +357,4 @@ TEST_F(WalletTxValidationTest, sign_verify_compact) {
     m_pubkeys.clear();
     ASSERT_FALSE(
         cbdc::transaction::validation::check_attestations(ctx, m_pubkeys, 2));
-}
-
-class MinterValidationTest : public ::testing::Test {
-  protected:
-    void SetUp() override {
-        const auto minter_pub = m_minter.generate_minter_key();
-        m_opts.m_minter_pubkeys.insert(minter_pub);
-    }
-
-    cbdc::transaction::wallet m_minter{};
-    cbdc::transaction::wallet m_not_minter{};
-    cbdc::config::options m_opts{};
-};
-
-// To Test:
-// * valid mint
-// * invalid mint (wrong signer)
-// * missing_output
-// * missing output value
-// * wrong number of witnesses
-// * wrong output committment
-TEST_F(MinterValidationTest, valid_mint) {
-    cbdc::transaction::full_tx tx = m_minter.mint_new_coins(5, 10);
-    auto err = cbdc::transaction::validation::check_tx(tx, m_opts);
-    ASSERT_FALSE(err.has_value());
-}
-
-TEST_F(MinterValidationTest, invalid_mint) {
-    cbdc::transaction::full_tx tx = m_not_minter.mint_new_coins(1, 1000);
-    auto err = cbdc::transaction::validation::check_mint_p2pk_witness(tx,
-                                                                      0,
-                                                                      m_opts);
-    ASSERT_TRUE(err.has_value());
-    ASSERT_EQ(
-        err.value(),
-        cbdc::transaction::validation::witness_error_code::invalid_minter_key);
-}
-
-TEST_F(MinterValidationTest, no_outputs) {
-    cbdc::transaction::full_tx tx = m_minter.mint_new_coins(5, 10);
-    tx.m_outputs.clear();
-
-    auto err = cbdc::transaction::validation::check_tx(tx, m_opts);
-    ASSERT_TRUE(err.has_value());
-}
-
-TEST_F(MinterValidationTest, no_output_value) {
-    cbdc::transaction::full_tx tx = m_minter.mint_new_coins(5, 10);
-    tx.m_outputs[0].m_value = 0;
-
-    auto err = cbdc::transaction::validation::check_tx(tx, m_opts);
-    ASSERT_TRUE(err.has_value());
-}
-
-TEST_F(MinterValidationTest, missing_witness) {
-    cbdc::transaction::full_tx tx = m_minter.mint_new_coins(5, 10);
-    tx.m_witness.clear();
-
-    auto err = cbdc::transaction::validation::check_tx(tx, m_opts);
-    ASSERT_TRUE(err.has_value());
-}
-
-TEST_F(MinterValidationTest, bad_witness_committment) {
-    cbdc::transaction::full_tx tx = m_minter.mint_new_coins(5, 10);
-    tx.m_outputs[0].m_witness_program_commitment = cbdc::hash_t{0};
-
-    auto err = cbdc::transaction::validation::check_tx(tx, m_opts);
-    ASSERT_TRUE(err.has_value());
 }
